# Faster equality in generic contexts