--- a/tests/AheadOfTime/Trimming/check.cmd
+++ b/tests/AheadOfTime/Trimming/check.cmd
@@ -1,2 +1,0 @@
-@echo off
-powershell -ExecutionPolicy ByPass -NoProfile -command "& """%~dp0check.ps1""""