--- a/tests/AheadOfTime/check.cmd
+++ b/tests/AheadOfTime/check.cmd
@@ -1,0 +1,3 @@
+@echo off
+powershell -ExecutionPolicy ByPass -NoProfile -command "& """%~dp0Trimming\check.ps1""""
+powershell -ExecutionPolicy ByPass -NoProfile -command "& """%~dp0Equality\check.ps1""""