… azure-pipelines.yml
… docs/release-notes/.FSharp.Core/8.0.300.md
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/FSharpCoreFunctions.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/Floats.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/Misc.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/OptionsAndCo.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/Structs.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/Tuples.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/Equality/ValueTypes.fs
… tests/benchmarks/CompiledCodeBenchmarks/MicroPerf/MicroPerf.fsproj