**What is this?**

This is the first part of [this effort](https://github.com/dotnet/fsharp/issues/16125) to resurrect the awesome work on compiler performance by @manofstick.

**How is this different?**

Many things have changed, among other things we now have a bigger team & community to review things and a different release cadence allowing us to dogfood things a bit.

More importantly, the original PRs often contained multiple ideas (fixes, breaking, optimizations, experiments, ...), whereas I want to narrow PRs as much as possible to make things easier to review and control.

**Is this PR breaking?**

A tiny bit. This is mostly an optimization.

Here is an example of code where the behavior does change:
```fsharp
[<Struct; CustomEquality; NoComparison>]
type Blah =
    interface IEquatable<Blah> with
        member lhs.Equals rhs = failwith "bang"

let eq x y = 
    printfn "hello" // long code insuring no inlining
    ...
    x = y // generic equality context

let result = eq (Blah()) (Blah())  // currently false, will be exception
```
We agreed that it's OK [here](https://github.com/dotnet/fsharp/pull/16615#issuecomment-1947440376).

**What's the source of inspiration here?**

This is essentially part of the #5112, with the following changes:
- Optimizer is not touched (there wasn't a consensus on that)
- tail calls are not touched (same, also partially was there to work around JIT issues that were resolved later)
- => hence IL is not changed
- some comparison optimizations are removed (I wasn't convinced by them)
- refactoring is removed (might be done in a followup)
- => hence diff is minimized
- benchmarks are added

**So where does this improve things?**

More theory and motivation is in [this document](https://github.com/fsharp/fslang-design/pull/765) - link will be updated once the PR merged.

*TL;DR: this improves things when `HashIdentity.Structural<'T>` comparison is used in non-inlined code.*

**Example optimization**

Let's look at the following code:

```fsharp
type Musician = {
    Name: string
    Surname: string
}

let musicians = [ 
    { Name = "Dave"; Surname = "Gahan" }
    { Name = "Jim"; Surname = "Morrisson" }
    { Name = "Robert"; Surname = "Smith" }
    { Name = "Dave"; Surname = "Grohl" }
    { Name = "Johnny"; Surname = "Marr" }
    { Name = "David"; Surname = "Gilmour" }
]

let getInitials musician =
    struct
        (musician.Name     |> Seq.head, 
         musician.Surname |> Seq.head)    

let result = 
    musicians 
    |> List.map getInitials
    |> List.distinct
```

---

How does `List.distinct` work?

There are 3 important parts of the algorithm to talk about.

1) Pick the comparer

The comparer is something implementing `IEqualityComparer<'T>` which means 2 methods: `GetHashCode(x)` and `Equals(x, y)`.

`List.distinct list` calls `List.distinctWithComparer HashIdentity.Structural<'T>` and the logic for picking `HashIdentity.Structural<'T>` (the comparer) is written in `prim-types.fs`.

2) Initialize the hash set for the distinct elements

This means calling `let hashSet = HashSet<'T>(comparer)` with the comparer picked above.

3) Add the elements to the hash set

So in our case,

```fsharp
hashSet.Add('D', 'G')
hashSet.Add('J', 'M')
hashSet.Add('R', 'S')
hashSet.Add('D', 'G')
hashSet.Add('J', 'M')
hashSet.Add('D', 'G')
```

---

Now, the `Add(element)` operation works in the following way:

1) Do all the bucket initialization stuff they teach about in the universities
2) Execute **GetHashCode** to get the hash of the element
3) Check if this hash is already present
4) If yes, execute **Equals** to see if this is the same element and decide on adding it to the hash set
5) Otherwise just add the element to the hash set

This means, in our case there will be:
- 6 `GetHashCode` calls (since there are 6 elements altogether)
- 3 `Equals` calls (since there are only 3 unique elements)

---

Now, what changes in this PR is that we become smarter at picking the faster comparer (step 1). This brings enormous benefit at doing all the things in the step 3.

Before, this is how the comparer picking would be executed:

```fsharp
List.distinct list
  List.distinctWithComparer HashIdentity.Structural<'T> list
    // check if this is a basic F# type - we optimize things for them
    FastGenericEqualityComparerTable<'T>.Function
      // no, this is not a basic type hence go the worst case - create generic equality comparer
      MakeGenericEqualityComparer<'T>
        // this is what it creates - we'll get to the consequences later
        { new IEqualityComparer<'T> with 
              member _.GetHashCode(x) = GenericHash x 
              member _.Equals(x,y) = GenericEquality x y }
```

Now, this is what is going on:

```fsharp
List.distinct list
  List.distinctWithComparer HashIdentity.Structural<'T> list
    // call "smart" `canUseDefaultEqualityComparer` to see if this is a type applicable for the default equality comparison
    FastGenericEqualityComparerTable<'T>.Function
      // yes it is
      EqualityComparer<'T>.Default
        // this is property but it basically creates kind of a "native" comparer for this type, like
        { new IEqualityComparer<(char, char)> with 
              member _.GetHashCode(x) = x.GetHashCode() 
              member _.Equals(x,y) = x.Equals(y) }
```

---

Hence, this is the difference for **each** of the 6 `GetHashCode` calls in question (taking first element ('D', 'G') as an example).

Before, using the generic equality comparer:

```fsharp
GetHashCode ('D', 'G')
  GenericHashIntrinsic ('D', 'G')
    GenericHashParamObj (object ('D', 'G'))                      // boxing!
      (IStructuralEquatable (ValueTuple ('D', 'G'))).GetHashCode()
        GetHashCodeCore ('D', 'G')
          comparer.GetHashCode('D')
            GenericHashParamObj (object ('D'))                    // boxing!
              (char 'D').GetHashCode()
          comparer.GetHashCode('G')
            GenericHashParamObj (object ('G'))                    // boxing!
              (char 'G').GetHashCode()
          HashCode.Combine
```

Now, using the "native" comparer:

```fsharp
GetHashCode ('D', 'G')
      (ValueTuple ('D', 'G').GetHashCode()
        GetHashCodeCore ('D', 'G')
          comparer.GetHashCode('D')
            (char 'D').GetHashCode()
          comparer.GetHashCode('G')
	    (char 'G').GetHashCode()
          HashCode.Combine
```

Now, this difference for **each** of the 3 `Equals` calls in question (taking the elements ('D', 'G') as an example).

Before, using the generic equality comparer:

```fsharp
Equals ('D', 'G') ('D', 'G')
  GenericEqualityIntrinsic ('D', 'G') ('D', 'G')
    GenericEqualityObj (object ('D', 'G')) (object ('D', 'G'))                               // boxing!
      (IStructuralEquatable (ValueTuple ('D', 'G'))).Equals(ValueTuple ('D', 'G'))
        comparer.Equals('D', 'D'))
          GenericEqualityObj (object ('D')) (object ('D'))                                         // boxing!
	    (char 'D').Equals(object 'D')
	      'D' == (char)'D'
        comparer.Equals('G', 'G'))
          GenericEqualityObj (object ('G')) (object ('G'))                                         // boxing!
	    (char 'G').Equals(object 'G')
	      'G' == (char)'G'
```

Now, using the "native" comparer:

```fsharp
Equals ('D', 'G') ('D', 'G')
      (ValueTuple ('D', 'G')).Equals(ValueTuple ('D', 'G'))
	      (char 'D').Equals(char 'D')
	        'D' == 'D'
	      (char 'G').Equals(char 'G')
	        'G' == 'G'
```

---

We can see that here we remove all the boxing and use the optimal call chain. This brings huge benefits for the tiny cost of executing the "smart" function on deciding about the comparer **once**.

The benefits of this approach vary based on the concrete algorithm and the elements in question. See the details about improved call chains in the spec mentioned above.

If we modify the example to have 1000 elements with 10 unique ones, we get the following results:

Before:
|       Method |     Mean |   Error |   StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|--------:|---------:|--------:|------:|------:|----------:|
| TheBenchmark | 142.4 us | 3.75 us | 10.82 us | 34.1797 |     - |     - | 210.49 KB |

After:
|       Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| TheBenchmark | 24.90 us | 0.871 us | 2.569 us | 0.1526 |     - |     - |     984 B |

Which means 6x faster and 213x less memory.

**Benchmarks**

Main targets: structs, enums, floats, and specia/l generic types:

<details>

<summary>Structs and enums</summary>

Before:
|       Method |     Mean |    Error |    StdDev |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|----------:|--------:|--------:|------:|----------:|
| FSharpStruct | 503.7 us | 36.91 us | 107.67 us | 68.3594 |       - |     - | 420.61 KB |
|   FSharpEnum | 142.7 us |  4.76 us |  13.73 us | 22.9492 |       - |     - | 140.65 KB |
| CSharpStruct | 148.2 us |  3.81 us |  10.94 us | 38.4521 | 12.8174 |     - | 237.58 KB |
|   CSharpEnum | 134.0 us |  8.43 us |  24.44 us | 22.8271 |       - |     - | 140.59 KB |

After:
|       Method |     Mean |    Error |    StdDev |   Median |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------- |---------:|---------:|----------:|---------:|--------:|-------:|------:|----------:|
| FSharpStruct | 73.44 us | 5.158 us | 15.046 us | 68.59 us |  0.1221 |      - |     - |     792 B |
|   FSharpEnum | 16.10 us | 0.396 us |  1.163 us | 16.35 us |  0.0458 |      - |     - |     336 B |
| CSharpStruct | 77.63 us | 2.285 us |  6.482 us | 79.15 us | 28.5645 | 9.4604 |     - |  179312 B |
|   CSharpEnum | 15.92 us | 0.638 us |  1.839 us | 16.21 us |  0.0916 |      - |     - |     656 B |

</details>

Huge improvements in both execution time and allocs, with especially remarkable results for native F# constructs.

<details>

<summary>Value tuples</summary>

Before:

|      Method |       Mean | Ratio |    Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------ |-----------:|------:|---------:|--------:|--------:|----------:|
| ValueTuple3 |   673.4 us |  1.00 |  61.5234 | 15.8691 |       - | 378.13 KB |
| ValueTuple4 |   812.2 us |  1.22 |  69.0918 | 19.7754 |       - | 424.98 KB |
| ValueTuple5 | 1,004.2 us |  1.50 |  84.9609 | 24.4141 |       - | 523.63 KB |
| ValueTuple6 | 1,100.7 us |  1.65 |  92.7734 | 23.4375 |       - | 570.48 KB |
| ValueTuple7 | 1,324.9 us |  1.97 | 117.1875 | 57.6172 | 29.2969 | 669.14 KB |
| ValueTuple8 | 1,461.9 us |  2.20 | 117.1875 | 58.1055 | 29.2969 | 762.85 KB |

After:

|      Method |     Mean | Ratio |   Gen 0 |   Gen 1 |   Gen 2 | Allocated |
|------------ |---------:|------:|--------:|--------:|--------:|----------:|
| ValueTuple3 | 173.0 us |  1.00 | 28.5645 |  9.3994 |       - | 175.11 KB |
| ValueTuple4 | 174.9 us |  1.03 | 28.5645 |  9.4604 |       - | 175.11 KB |
| ValueTuple5 | 208.9 us |  1.22 | 34.4238 | 11.3525 |       - | 211.29 KB |
| ValueTuple6 | 217.0 us |  1.26 | 34.4238 | 11.3525 |       - | 211.29 KB |
| ValueTuple7 | 293.7 us |  1.73 | 29.2969 | 29.2969 | 29.2969 | 247.48 KB |
| ValueTuple8 | 293.8 us |  1.73 | 29.2969 | 29.2969 | 29.2969 | 247.48 KB |

</details>

~80% in speed and ~50% in memory reduction, also much steeper ratios' increase for both.

<details>

<summary>Options and co</summary>

Before:

|      Method |     Mean |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|--------:|------:|----------:|
|      Option | 165.0 us | 16.3574 |  3.1738 |     - | 101.74 KB |
| ValueOption | 157.1 us | 28.8086 |  4.3945 |     - | 177.02 KB |
|      Result | 186.3 us | 40.4053 | 10.0098 |     - | 248.25 KB |

After:

|      Method |     Mean |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|-------:|------:|----------:|
|      Option | 82.13 us | 12.6953 | 3.0518 |     - |  78.33 KB |
| ValueOption | 55.09 us |  9.7656 | 1.5869 |     - |  59.98 KB |
|      Result | 75.92 us | 22.5830 | 5.6152 |     - | 138.93 KB |

</details>

50-75% speed and 25-75% memory improvements.

<details>

<summary>Nullable<'T></summary>

Before:

|      Method |     Mean |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|--------:|------:|----------:|
|    Nullable | 443.7 us | 24.9023 |  3.9063 |     - | 153.66 KB |

After:

|      Method |     Mean |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|-------:|------:|----------:|
|    Nullable | 60.16 us |  9.7656 | 1.5869 |     - |  59.94 KB |

</details>

About 7x speed and 3x memory improvements.

<details>

<summary>Floats</summary>

Before:

|     Method |      Mean |    Error |    StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|----------- |----------:|---------:|----------:|--------:|-------:|------:|----------:|
|    FloatER |  51.47 us | 2.228 us |  6.570 us |  7.0801 | 0.3662 |     - |  43.68 KB |
|  Float32ER |  56.39 us | 2.566 us |  7.525 us |  7.1106 | 0.3662 |     - |  43.68 KB |
|   FloatPER | 149.49 us | 2.952 us |  7.513 us | 38.3301 | 2.6855 |     - | 231.34 KB |
| Float32PER | 136.09 us | 5.584 us | 16.201 us | 37.5977 | 2.4414 |     - | 227.01 KB |

After:

|     Method |     Mean |    Error |    StdDev |   Median |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|----------- |---------:|---------:|----------:|---------:|--------:|-------:|------:|----------:|
|    FloatER | 15.26 us | 0.487 us |  1.397 us | 15.38 us |  3.2806 | 0.1678 |     - |   20.1 KB |
|  Float32ER | 15.80 us | 0.316 us |  0.666 us | 15.82 us |  3.2806 | 0.1678 |     - |   20.1 KB |
|   FloatPER | 82.88 us | 5.580 us | 16.452 us | 93.36 us | 14.9536 | 1.2817 |     - |  90.46 KB |
| Float32PER | 94.22 us | 1.851 us |  3.904 us | 93.97 us | 14.1602 | 0.9766 |     - |  86.43 KB |

</details>

PER comparison still takes more time and memory but still the improvements are 2-3 times in all cases.

---

Also (positively) affected: basic types, arrays, reference types - due to shorter call chains and less casting:

<details>

<summary>Arrays</summary>

Before:

| Method |       Mean |    Error |   StdDev |
|------- |-----------:|---------:|---------:|
|  Int32 |   974.3 us | 73.48 us | 214.3 us |
|  Int64 | 1,090.7 us | 58.65 us | 172.9 us |
|   Byte | 1,075.3 us | 41.56 us | 121.9 us |
|    Obj | 1,451.8 us | 43.91 us | 128.8 us |

After:

| Method |     Mean |    Error |   StdDev |
|------- |---------:|---------:|---------:|
|  Int32 | 253.3 us | 18.09 us | 52.78 us |
|  Int64 | 312.3 us | 14.06 us | 41.45 us |
|   Byte | 246.5 us |  6.11 us | 17.82 us |
|    Obj | 489.3 us | 17.69 us | 51.61 us |

</details>

About 3x faster.

<details>

<summary>F# basic types</summary>

Before (`countBy`):
|  Method |      Mean |     Error |    StdDev |
|-------- |----------:|----------:|----------:|
|    Bool |  39.06 us |  1.936 us |  5.709 us |
|   SByte |  55.23 us |  2.032 us |  5.992 us |
|    Byte |  50.62 us |  1.617 us |  4.766 us |
|   Int16 |  85.46 us |  4.668 us | 13.764 us |
|  UInt16 |  86.66 us |  3.189 us |  9.351 us |
|   Int32 |  86.21 us |  4.690 us | 13.827 us |
|  UInt32 |  87.69 us |  4.911 us | 14.480 us |
|   Int64 | 112.81 us |  3.962 us | 11.681 us |
|  UInt64 | 112.66 us |  4.003 us | 11.550 us |
|  IntPtr | 114.61 us |  3.430 us | 10.114 us |
| UIntPtr | 109.40 us |  3.322 us |  9.796 us |
|    Char |  98.99 us |  2.825 us |  8.330 us |
|  String | 214.52 us |  5.968 us | 17.503 us |
| **Decimal** | 315.84 us | 12.545 us | 36.988 us |

After (`countBy`):

|  Method |      Mean |    Error |    StdDev |
|-------- |----------:|---------:|----------:|
|    Bool |  29.19 us | 1.608 us |  4.740 us |
|   SByte |  50.72 us | 2.167 us |  6.390 us |
|    Byte |  47.40 us | 1.719 us |  5.069 us |
|   Int16 |  83.50 us | 3.508 us | 10.342 us |
|  UInt16 |  84.48 us | 2.949 us |  8.649 us |
|   Int32 |  87.24 us | 2.670 us |  7.832 us |
|  UInt32 |  86.17 us | 3.630 us | 10.703 us |
|   Int64 |  95.78 us | 4.763 us | 14.044 us |
|  UInt64 | 113.86 us | 4.278 us | 12.479 us |
|  IntPtr | 110.92 us | 3.192 us |  9.412 us |
| UIntPtr | 105.11 us | 3.219 us |  9.440 us |
|    Char |  91.03 us | 4.164 us | 12.211 us |
|  String | 214.23 us | 5.859 us | 17.276 us |
| **Decimal** | 150.39 us | 3.703 us | 10.683 us |

Before (`distinct`):

|  Method |      Mean |     Error |    StdDev |
|-------- |----------:|----------:|----------:|
|    Bool |  9.489 us | 0.7345 us |  2.142 us |
|   SByte | 12.568 us | 0.4049 us |  1.168 us |
|    Byte | 12.393 us | 0.4176 us |  1.231 us |
|   Int16 | 23.539 us | 0.8906 us |  2.626 us |
|  UInt16 | 22.311 us | 0.8351 us |  2.449 us |
|   Int32 | 22.302 us | 0.4448 us |  1.180 us |
|  UInt32 | 22.092 us | 0.4760 us |  1.319 us |
|   Int64 | 25.567 us | 0.8679 us |  2.518 us |
|  UInt64 | 26.200 us | 1.4150 us |  4.172 us |
|  IntPtr | 25.528 us | 1.4250 us |  4.202 us |
| UIntPtr | 25.000 us | 0.8785 us |  2.590 us |
|    Char | 22.883 us | 0.8124 us |  2.383 us |
|  String | 47.659 us | 1.6451 us |  4.799 us |
| **Decimal** | 58.086 us | 4.7662 us | 13.828 us |

After:

|  Method |      Mean |     Error |   StdDev |
|-------- |----------:|----------:|---------:|
|    Bool |  9.408 us | 0.9121 us | 2.689 us |
|   SByte | 11.795 us | 0.3516 us | 1.020 us |
|    Byte | 11.225 us | 0.4078 us | 1.170 us |
|   Int16 | 23.211 us | 0.7982 us | 2.354 us |
|  UInt16 | 21.682 us | 1.0798 us | 3.184 us |
|   Int32 | 20.035 us | 0.6128 us | 1.787 us |
|  UInt32 | 20.123 us | 0.8883 us | 2.591 us |
|   Int64 | 22.967 us | 0.7908 us | 2.319 us |
|  UInt64 | 24.757 us | 1.3208 us | 3.832 us |
|  IntPtr | 26.448 us | 1.3528 us | 3.989 us |
| UIntPtr | 24.596 us | 1.1292 us | 3.330 us |
|    Char | 22.670 us | 0.8961 us | 2.642 us |
|  String | 40.480 us | 1.3767 us | 3.994 us |
| **Decimal** | 35.401 us | 1.4948 us | 4.289 us |

Note that decimals also show improvement in memory allocations:

Before (`countBy`)

|  Method |     Mean |    Error |   StdDev |   Median |   Gen 0 |   Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|---------:|--------:|--------:|------:|----------:|
| Decimal | 149.9 us | 13.45 us | 39.01 us | 136.1 us | 38.4521 | 12.8174 |     - | 237.55 KB |

After (`countBy`)

|  Method |     Mean |    Error |    StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|----------:|--------:|-------:|------:|----------:|
| Decimal | 65.17 us | 5.013 us | 13.807 us | 28.5645 | 9.4604 |     - | 175.09 KB |

Before (`distinct`)

|  Method |     Mean |    Error |    StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|----------:|--------:|-------:|------:|----------:|
| Decimal | 57.83 us | 6.097 us | 17.977 us | 26.3062 | 6.5308 |     - | 162.35 KB |

After (`distinct`)

|  Method |     Mean |    Error |   StdDev |   Gen 0 |  Gen 1 | Gen 2 | Allocated |
|-------- |---------:|---------:|---------:|--------:|-------:|------:|----------:|
| Decimal | 34.79 us | 1.572 us | 4.585 us | 21.2708 | 5.3101 |     - |  131.1 KB |

</details>

These mostly stay the same (as expected), apart from `decimal`, which show ~50% speed and ~25% alloc improvements. 

<details>

<summary>Records</summary>

Before:
|               Method |     Mean |   Error |   StdDev |
|--------------------- |---------:|--------:|---------:|
…