@vzarytovskii 2024-02-08
Benchmark results are weird, numbers before and after are within the statistical error. What about allocations? We shouldn't be boxing as much now?