@vzarytovskii 2024-02-14
Since it touches an important part of the language, I would like to have three things:

1. What changed, in the form of
   Given (generic) code: ...
   How it worked before: ...
   How does it work now: ...

2. Some AOT compilation tests involving changed equality. It can be a new project, like the trimming one we have. We should start having more AOT testing.

3. Can we have some compiler perf comparison/profiling of using compiler+fslib before and after changes on something like FCS?