@dsyme 2024-02-14
> Well... It'll be calling IEquatable.Equals<>, not object.Equals, which is hence a breaking change - albeit I feel in a fashion that should be part of an evolution...

@manofstick Could you remind me of the specific thing that causes the call to `IEquatable.Equals<`, and which types it applies to.  I assume it comes from ` EqualityComparer<'T>.Default` and thus only applies to types that pass `canUseDefaultEqualityComparer`?