@psfinaki 2024-02-14
@vzarytovskii I think it's a good idea to have that doc (from the other PR) in the design repo. This PR is meant to be an optimization and hence focuses on the implementation. I am looking at the spec while working on this and my impression _so far_ is that the spec doesn't go _that deep_ in the implementation to make this change anyhow misalign with it. If we discover anything of that nature, we can make the RFC or further discuss it.