@psfinaki 2024-02-14
I added benchmarks for the value tuples, options and that stuff - the results are very convincing so far. 
More to come soon!