@psfinaki 2024-02-15
@manofstick just to be sure we're on the same page - which input cast do you mean? :)