@manofstick 2024-02-15
@psfinaki 

...forgive my github comment coding, I think memory serves, but it's more than enough for the gist of things if it's incorrect...

```fsharp
type Blah(...) =
    ...

    override lhs.Equals (rhsObj:obj) = // (1)
        if obj.ReferenceEquals (lhs, rhsObj) then
            true
        else 
            match rhs with
            | :? Blah as rhs -> lhs.Equals rhs // (2)
            | _ -> false

    interface IEquatable<Blah> with
        member lhs.Equals rhs = // (3)
            // actually do the check
```

Well `(1)` would of been the entry point, where at `(2)` would of been the cast (well type check), but `EqualityComparer<Blah>.Default` will call `(3)` directly.

Most likely the actually equality check in these cases will be the significant component of the cost, so savings is minimal, but as mentioned, this is just for complete understanding of what's going on here...