@dsyme 2024-02-15
@psfinaki I believe it should be possible to write a test case that detects the change, e.g.
```fsharp
[<Sealed, Struct>]
type Blah() =
    override lhs.Equals (rhsObj:obj) = true
    interface IEquatable<Blah> with
        member lhs.Equals rhs = failwith "bang"

let eq x y = 
    printfn "hello" // do this enough times to make sure no inlining
    printfn "hello" // do this enough times to make sure no inlining
    printfn "hello" // do this enough times to make sure no inlining
    printfn "hello" // do this enough times to make sure no inlining
    (x = y) // generic equality context

eq (Blah()) (Blah())  // true prior to this, exception now
```