@psfinaki 2024-02-16
Yes, correct. For reference, I guess this is the minimal code to demo the behavior difference:

```fsharp
[<Struct; CustomEquality; NoComparison>]
type Blah =
	interface IEquatable<Blah> with
		member lhs.Equals rhs = failwith "bang"

let eq x y = 
	printfn "hello" // do this enough times to make sure no inlining
	...
	x = y // generic equality context

let result = eq (Blah()) (Blah())  // true prior to this, exception now
```

Alright, so we've identified what breaks here and blessed it at the same time. Good stuff, I will update the spec and the description to reflect this.